feat: Prediction inequality & adjust-alpha
Design the prediction inequality as follows:

DirtyPageNumber*AverageDirtyByteCountPerPage/TR + RunTime + alpha
<= 0.94*TL

where TR is the transfer rate, TL is the target latency, RunTime
is the run stages length so far and alpha is the copensating value
for snapshot stages, signals between threads and other delays.
The factor, 0.94, is a reasonable ratio to approach the
target latency. A DirtyPageNumber counter is created and
incremented whenever a new page is dirtied.
AverageDirtyByteCountPerPage is predicted from a measurement made
at an earlier point of the run stage. When a new page is dirtied
or when one of a set of pre-defined timers expires, the above
inequality is checked, and the run stage is terminated if the
inequality does not hold.

We perform these checks at these pre-defined timers because
sometimes a master VM may not dirty any new pages later in a
run stage and these timer events ensure that the master VM is
always capable of terminating the current run stage even when
no new pages are dirtied.

Design the Auto-Adjustment of Alpha Time:

alpha is used to compensate micellaneous time such as the time
spent by communication between master and slave VM, the time spent
by snapshot stages, the time for context switches between
QEMU threads, the time spent by slave VM to send ACKs, and other
time-consuming sources. alpha is adjusted in runtime automatically
by retrospecting historical latencies. The adjust-alpha algorithm
in each periodic adjustment (every 1000 rounds) will update alpha.
According to the prediction inequality, it is obvious that if the
average latency is larger than target latency, we should increase
the alpha by the distance between average latency and target latency.
On the other side, when the average latency is smaller than the
approached target latency, then this alpha is too high and decrease
by the distance between average latency and approached target latency.
In this way, this alpha value is not fixed to an arbitrary value and
it can adjust itself according to the platform situation.